+++ tests/testthat/test_misc.R
@@ -185,6 +185,51 @@ test_that('set_engine works as a generic', {
 test_that('check_for_newdata points out correct context', {
   fn <- function(...) {check_for_newdata(...); invisible()}
   expect_snapshot(error = TRUE,
-    fn(newdata = "boop!")
+                  fn(newdata = "boop!")
+  )
+})
+
+test_that('check_outcome works as expected', {
+  reg_spec <- linear_reg()
+
+  expect_no_error(
+    check_outcome(1:2, reg_spec)
+  )
+
+  expect_no_error(
+    check_outcome(mtcars, reg_spec)
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    check_outcome(factor(1:2), reg_spec)
+  )
+
+  class_spec <- logistic_reg()
+
+  expect_no_error(
+    check_outcome(factor(1:2), class_spec)
+  )
+
+  expect_no_error(
+    check_outcome(lapply(mtcars, as.factor), class_spec)
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    check_outcome(1:2, class_spec)
+  )
+
+  # Fake specification to avoid having to load {censored}
+  cens_spec <- logistic_reg()
+  cens_spec$mode <- "censored regression"
+
+  expect_no_error(
+    check_outcome(survival::Surv(1, 1), cens_spec)
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    check_outcome(1:2, cens_spec)
   )
 })